refactor(auth): reorganize external browser sso modules
Move external browser SSO support under auth::external_browser to keep implementation details in one module tree.

Preserve the public API from lib.rs while updating internal references to the new layout.
